requirements.txt
streamlit~=1.51.0
dotenv~=0.9.9
google-generativeai
openmeteo-requests
requests-cache~=1.2.1
retry-requests
pytest~=8.4.2
numpy~=2.3.4
openmeteo_requests
protobuf~=5.29.5
python-dotenv~=1.2.1
matplotlib~=3.10.7
openmeteo_requests
requests~=2.32.5
openmeteo_requests~=1.7.4